Extract configurations, clients, completions
