+++ __init__.py
@@ -1,1 +1,2 @@
 from .core.main import *
+from .features.completion import *
