+++ core/documents.py
@@ -1,0 +1,26 @@
+import sublime
+
+from collections import OrderedDict
+
+try:
+    from typing import Any, List, Dict, Tuple, Callable, Optional
+    assert Any and List and Dict and Tuple and Callable and Optional
+except ImportError:
+    pass
+
+
+from .protocol import Point
+from .url import filename_to_uri
+
+
+def get_document_position(view: sublime.View, point) -> 'Optional[OrderedDict]':
+    file_name = view.file_name()
+    if file_name:
+        if not point:
+            point = view.sel()[0].begin()
+        d = OrderedDict()  # type: OrderedDict[str, Any]
+        d['textDocument'] = {"uri": filename_to_uri(file_name)}
+        d['position'] = Point.from_text_point(view, point).to_lsp()
+        return d
+    else:
+        return None
